Fix(client): mypage prerender error 해결
