+++ apps/client/src/components/MyPage/Content.tsx
@@ -1,0 +1,19 @@
+'use client';
+
+import { useSearchParams } from 'next/navigation';
+import ProfileSetting from './ProfileSetting';
+import QuizHistorySetting from './QuizHistorySetting';
+
+const Content = () => {
+  const searchParams = useSearchParams();
+  const tab = searchParams.get('tab');
+
+  return (
+    <div className='w-full max-w-[81rem]'>
+      {tab === 'profile' && <ProfileSetting />}
+      {tab === 'my-quizzes' && <QuizHistorySetting />}
+    </div>
+  );
+};
+
+export default Content;
